Touchscreen support and beginnings of blur overlay
Creating a blurry overlay with aperture is going to be hard. SVG filters don't offer a way to clip off the effect unless you apply to an image -- not possible with general purpose website tool. For now we have to stick with opaque overlays.

Added support for touch-events on your phones and ipads. But auto-scrolling might be an issue here, as overlay does not update and we get a preview underneath. Need to test on actual devices, perhaps window.innerwidth etc work differently, as this is not an issue on desktop devics