fixing the html2canvas scaling issue
Some browsers generate a scaled down or up screenshot. Previously we had assumed a multiplier of 2, however this is only true on some browser/OS combinations. We now explicitly calculate the ration of the canvas style dimensions (the actual ones) with the resolution of the screen shot, and use this as a multiplier, instead of assuming