overlay callback function and percentage of screen option
Added the option to put a percentage in as an aperture width, which maps to percentage of screen height

Also added the option to specify a function that would be run on the succesful generation of a blurred overlay (as this can take up to a second). This can be used to hide the screen for instance, until the blur is generated.